marvel_bins.py: Uncomment iteration setters.
Not sure why these lines were commented out - I was able to trace some odd results I was getting to those lines.

Also, unrelatedly, I found that marvel_bins.py did not work because a folder name was being given as the number of threads to prokka and hmmscan - this can be reproduced by just running it on some data.

Also, another small thing, hmmscan is not listed as a dependency in the README.

Hope that all helps. How come the "issues" tab on GitHub has been disabled? I almost did not submit the above comments because there was no obvious way to do so.

Thanks, ben